Safer way to retrieve a request’s path when verifying response signature
When using a proxy, the attribute `http.request` is of type `BasicHttpRequest` and can’t be cast to `HttpRequestWrapper`.
In order to access the request’s path in a response interceptor we rather explicitly set it on the context using our own attribute in a request interceptor.
Upgraded a few dependencies and fixed an issue where the `EventLogger` could be `null`.