Even more robust response resource handling
Can this be more goldplated? I dare anyone! 🥇

Basically, there are some challenges with offering the InputStream of
the HTTP response entity. I the case of the third party getting the
InputStream, she/he is expected to do proper closing of this stream, which
will propagate to close any resources of the response itself. But in the
case of exceptions happening during the acquisition of this stream, it
is important that we make sure there are no resource leaks, as the third
party does not receive any means to do the closing by themselves.