A substantial restructuring commit
Removing a large amount of "orphaned" functionality, things that don't
actually work (nor is necessary) with the version 7 of the API, but have
been left since earlier versions of the API.

Splitting up the internal interfaces of the client for better
comprehension: MessageDeliveryApi, DocumentApi, and InboxApi.

Various other minor cleanups.