Change how we get archive documents by uuid and referenceid. Add api for adding a new unique uuid to archiveDocuments
To be able to get urls from EntryPoint, I added the possibility
to get SenderId-based EntryPoint. This has been a feature in Digipost api
for several years, but never been implemented in the java-client.
By adding a cache og entrypoint pr senderId in the client and using
the entrypoint, we can get rid of hardcoded urls for the archive
functionality.

I also changed how we set referenceid and deletion date for ArchiveDocuments.
These are optional fields and should not be added by constructor imho.
A builder pattern makes the use much cleaner.
I also added TTL-style retention periode for the ArchiveDocument.

New:
* Get document by uuid
* Add new unique uuid to document