Warn if you add more filer/attachments to a message without having them specified in the actual message
The case this tries to address is where you have some type of looping to
add attachments, or maby the missing of this looping. It might be easy
to forget to go from a non looping code to a looping code that you
need to add all the attachments and also all the actual files.