Method to get list of tags for user, using personal identification number
Using this feature requires the ADMINISTER_TAGS Digipost API permission.
It can be used to list all current tags for a user, which as for now only
covers whether the user receives public mail or not. Filter the resulting
list using the TagType enum to derive desired user properties. This
functionality is as of yet primarily used for Digdir/KRR to manage
public mail recipients.