Don't remove digipost id from print messages
This fixes a bug where all ids (digipost address, pin, name and addres
and organisation number) gets removed before sending to the api. The
result was that the API got a direct print message, even if it was
originally a print fallback message, thus returning errors if the sender
was not allowed to use direct print.

(Message.copyMessageWithOnlyPrintDetails,
MessageDeliverer.setMapAndMessageToPrint) Copy all fields of the input
message (was: omitting ids and a few other fields). Rename to
copyPrintMessage, since it no longer narrows the result to print
details. Doc fix.

Remove obsolete validation of single channel in the recipient
(was: contradicts ability of print fallback in the API, that is, two
channels).