Error code for duplicate document in same batch
It can be useful to distinguish between this new
error code and the existing DUPLICATE_DOCUMENT_ID
error code when using the batch functionality of
the Digipost API. This new error code is a special
case of duplicate uuid for which the duplicate
is part of the same batch, and has the same document
hash, as the message which generated the error.
In such cases, it is safe for the sender to assume
the message has already been submitted as part of
the batch. This requires that the document hash
property is specified correctly.